split all logging tests into a separate file
